+++ app/src/main/java/org/simple/clinic/home/patients/PatientsTabViewEffectHandler.kt
@@ -1,0 +1,20 @@
+package org.simple.clinic.home.patients
+
+import org.simple.clinic.mobius.ViewEffectsHandler
+
+class PatientsTabViewEffectHandler(
+    private val uiActions: PatientsTabUiActions
+) : ViewEffectsHandler<PatientsTabViewEffect> {
+
+  override fun handle(viewEffect: PatientsTabViewEffect) {
+    when (viewEffect) {
+      OpenEnterOtpScreen -> uiActions.openEnterCodeManuallyScreen()
+      is OpenPatientSearchScreen -> uiActions.openPatientSearchScreen(viewEffect.additionalIdentifier)
+      ShowUserWasApproved -> uiActions.showUserStatusAsApproved()
+      HideUserAccountStatus -> uiActions.hideUserAccountStatus()
+      OpenScanBpPassportScreen -> uiActions.openScanSimpleIdCardScreen()
+      OpenTrainingVideo -> uiActions.openYouTubeLinkForSimpleVideo()
+      ShowAppUpdateAvailable -> uiActions.showAppUpdateDialog()
+    }
+  }
+}